checkpoint: renames "TextToImage/types" to "TextToImage/Pipeline"
