+++ src/features/TextToImage/Pipeline/Output/definition.ts
