+++ src/features/TextToImage/exports.ts
@@ -2,4 +2,4 @@ export * as Client from './client';
 
 export * as Server from './Server';
 
-export type * from './types';
+export type * from './Pipeline';
